// comfyui/ui/src/engine/particles/types.ts
export type ModulationCurve =
  | { type: "constant"; value: number }
  | { type: "linear"; start: number; end: number }
  | {
      type: "curve";
      points: Array<{
        time: number;
        value: number;
        inTangent?: number;
        outTangent?: number;
      }>;
    }
  | { type: "random"; min: number; max: number }
  | {
      type: "randomCurve";
      minCurve: ModulationCurve;
      maxCurve: ModulationCurve;
    };